refactor(ipc): extract @openomni/ipc from coordinator (#496)
Move the worker-process IPC transport (NDJSON framing with 16 MiB cap,
bidirectional Unix-socket client/server, typed transport errors) out of
packages/coordinator/src/ipc/ into the standalone protocol-only package
@openomni/ipc, history-preservingly. Serializable schemas stay in
@openomni/protocol; the package is driver-band consumable as a published
contract and keeps reverse (server -> owner-device) connections.

- coordinator: imports @openomni/ipc, barrel re-exports no IPC surface;
  dep ratchet now enforces protocol + ipc only (source-level red/green
  proven with a synthetic session import)
- apps/server worker-entry and the openomni policy-echo harness import
  createIpcServer from @openomni/ipc directly (coordinator import gone)
- decodeMessage/IpcMessage de-exported (knip baseline shrinks by 2)
- gates: check-deps ipc ring rule, knip workspace entry, CI ipc test
  step, coverage baseline split (coordinator 1374->1045 lines, ipc 330)

Owner-approved early start (2026-08-10 re-order): #496 pulled ahead of
the #494->#495 chain because it is mechanically independent.
