+++ .github/workflows/ci.yml
@@ -103,6 +103,9 @@ jobs:
       - name: Test (protocol)
         run: bun test --timeout 15000 --coverage --coverage-reporter=lcov --coverage-dir=coverage
         working-directory: packages/protocol
+      - name: Test (ipc)
+        run: bun test --timeout 15000 --coverage --coverage-reporter=lcov --coverage-dir=coverage
+        working-directory: packages/ipc
       - name: Test (openomni)
         run: bun run test:ci
         working-directory: packages/openomni
