+++ knip.json
@@ -33,6 +33,10 @@
       "entry": ["src/index.ts", "test/**/*.test.ts", "test/harness/*.ts"],
       "project": ["src/**/*.ts", "test/**/*.ts"]
     },
+    "packages/ipc": {
+      "entry": ["src/index.ts", "test/**/*.test.ts"],
+      "project": ["src/**/*.ts", "test/**/*.ts"]
+    },
     "packages/protocol": {
       "entry": ["src/index.ts", "test/**/*.test.ts", "test/**/*.d.ts"],
       "project": ["src/**/*.ts", "test/**/*.ts", "test/**/*.d.ts"]
