--- packages/coordinator/src/ipc/index.ts
@@ -1,1 +1,0 @@
-export { createIpcServer } from "./server";
