+++ packages/ipc/package.json
@@ -1,16 +1,16 @@
 {
-  "name": "@openomni/coordinator",
+  "name": "@openomni/ipc",
   "version": "0.0.1",
   "type": "module",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "scripts": {
+    "build": "tsc -p tsconfig.build.json",
     "check-types": "tsc --noEmit",
     "lint": "bunx biome check ./src",
     "test": "bun test"
   },
   "dependencies": {
-    "@openomni/ipc": "workspace:*",
     "@openomni/protocol": "workspace:*"
   },
   "devDependencies": {
