+++ packages/ipc/src/errors.ts
