+++ packages/ipc/src/framing.ts
