+++ script/conformance/knip-baseline.json
@@ -6,7 +6,6 @@
     "exports apps/server/src/channel/discord/surface.ts sendDiscordMessage",
     "exports apps/server/src/ingress/bridge.ts buildResidentAgentDef",
     "exports apps/server/src/tool/mcp/provider-audit.ts readSessionId",
-    "exports packages/coordinator/src/ipc/server.ts decodeMessage",
     "exports packages/openomni/src/execution-runtime/tool/builtins/edit.ts EDIT_PROMPT",
     "exports packages/openomni/src/execution-runtime/tool/builtins/glob.ts GLOB_PROMPT",
     "exports packages/openomni/src/execution-runtime/tool/builtins/grep.ts GREP_PROMPT",
@@ -17,7 +16,6 @@
     "exports packages/session/src/bus-persistence/event-query.ts toEventRecord",
     "exports packages/session/src/storage/sqlite-part-adapter.ts getPartStartTime",
     "files packages/openomni/test/dispatch/policy-registration.contract-test.ts",
-    "types packages/coordinator/src/ipc/server.ts IpcMessage",
     "types packages/coordinator/src/worker-manager/index.ts DeliverTask",
     "types packages/coordinator/src/worker-manager/worker-pool.ts DeliverTask",
     "types packages/openomni/src/execution-runtime/child-agent/types.ts ChildStatus",
